pyproject.toml
[tool.poetry]
name = "test-task-for-doovertech"
version = "0.1.0"
description = ""
authors = ["Ernur Eskermes <eskermesernur86@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
Pillow = "^8.4.0"
djangorestframework-simplejwt = "^5.0.0"
braintree = "^4.12.0"
django-celery-beat = "^2.2.1"
celery = "^5.1.2"
redis = "^3.5.3"
drf-yasg = "^1.20.0"
psycopg2 = "^2.9.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
